改进

+++ src/utils/ReadConf.cpp
@@ -92,7 +92,7 @@ utils::ReadConf::ReadConf() {
                 cfg.minFruit = stoi(ret[1]);
             } else if (ret[0] == "heath") {
                 utils::trim(ret[1]);
-                cfg.heath = stol(ret[1]);
+                cfg.heath = stof(ret[1]);
             } else if (ret[0] == "death") {
                 utils::trim(ret[1]);
                 cfg.death = stoi(ret[1]);
